requirements.txt
line-bot-sdk == 3.7.0
flask == 3.0.0
selenium ==4.10.0
webdriver-manager == 4.0.2
beautifulsoup4 == 4.12.2
requests == 2.29.0